modify : 로그인이 필요한 회원수정, 회원탈퇴 기능에 기능을 사용하는 유저가 맞는지 확인하는 로직 추가[#2]

+++ src/main/java/me/moon/ticketReservation/supplier/service/SupplierService.java
@@ -1,6 +1,9 @@
 package me.moon.ticketReservation.supplier.service;
 
 import lombok.RequiredArgsConstructor;
+import me.moon.ticketReservation.common.exception.ErrorCode;
+import me.moon.ticketReservation.login.entity.SessionUser;
+import me.moon.ticketReservation.login.exception.UnauthorizedAccessException;
 import me.moon.ticketReservation.supplier.dto.SupplierResponseDto;
 import me.moon.ticketReservation.supplier.dto.SupplierSaveRequestDto;
 import me.moon.ticketReservation.supplier.dto.SupplierUpdateRequestDto;
@@ -26,11 +29,18 @@ public class SupplierService {
         return new SupplierResponseDto().of(supplier);
     }
 
-    public SupplierResponseDto update(String supplierId, SupplierUpdateRequestDto dto) {
+    public SupplierResponseDto update(SessionUser sessionUser, String supplierId, SupplierUpdateRequestDto dto) {
         Supplier supplier = supplierMapper.findById(supplierId);
+        if ( !compareUserAndSessionUser(sessionUser, supplier) ) {
+            throw new UnauthorizedAccessException(sessionUser.getEmail(), ErrorCode.UNAUTHORIZED_ACCESS);
+        }
         supplier.update(dto);
         supplierMapper.update(supplier);
         return new SupplierResponseDto().of(supplier);
+    }
+
+    private boolean compareUserAndSessionUser(SessionUser sessionUser, Supplier supplier) {
+        return sessionUser.getEmail().equals(supplier.getEmail());
     }
 
     private boolean isDuplicateEmail(String email) {
@@ -38,7 +48,11 @@ public class SupplierService {
     }
 
 
-    public void delete(String supplierId) {
+    public void delete(SessionUser sessionUser, String supplierId) {
+        Supplier supplier = supplierMapper.findById(supplierId);
+        if ( !compareUserAndSessionUser(sessionUser, supplier) ) {
+            throw new UnauthorizedAccessException(sessionUser.getEmail(), ErrorCode.UNAUTHORIZED_ACCESS);
+        }
         int result = supplierMapper.deleteById(supplierId);
         if (result != 1){
             throw new WithdrawalException(supplierId);
